Reorganized into a simulate submodule
