+++ qspectra/__init__.py
@@ -1,9 +1,8 @@
 from bath import DebyeBath, ArbitraryBath, UncoupledBath
 from constants import CM_FS_LINEAR, CM_FS, CM_K, GAUSSIAN_SD_FWHM
-from dynamical_models.liouville_space import (n_excitations, den_to_vec,
-                                              vec_to_den)
-from dynamical_models.redfield import RedfieldModel
-from dynamical_models.unitary import UnitaryModel
+from dynamics.liouville_space import n_excitations, den_to_vec, vec_to_den
+from dynamics.redfield import RedfieldModel
+from dynamics.unitary import UnitaryModel
 from hamiltonian import (Hamiltonian, ElectronicHamiltonian,
                          VibronicHamiltonian)
 from operator_tools import unit_vec, basis_transform, all_states
@@ -11,10 +10,10 @@ from polarization import (polarization_vector, invariant_weights_4th_order,
                           list_polarizations, FOURTH_ORDER_INVARIANTS,
                           MAGIC_ANGLE)
 from pulse import CustomPulse, GaussianPulse
-from simulate import (simulate_dynamics, simulate_with_fields, simulate_pump,
-                      linear_response, absorption_spectra, impulsive_probe)
-from utils import fourier_transform
-import utils
+from simulate.eom import simulate_dynamics, simulate_with_fields, simulate_pump
+from simulate.response import (linear_response, absorption_spectra,
+                               impulsive_probe)
+from simulate.utils import fourier_transform, integrate
 
 __all__ = ['DebyeBath', 'ArbitraryBath', 'UncoupledBath', 'CM_FS_LINEAR',
            'CM_FS', 'CM_K', 'GAUSSIAN_SD_FWHM', 'ElectronicHamiltonian',
@@ -24,4 +23,4 @@ __all__ = ['DebyeBath', 'ArbitraryBath', 'UncoupledBath', 'CM_FS_LINEAR',
            'CustomPulse', 'GaussianPulse', 'RedfieldModel', 'UnitaryModel',
            'simulate_dynamics', 'simulate_with_fields', 'simulate_pump',
            'linear_response', 'absorption_spectra', 'impulsive_probe',
-           'fourier_transform', 'utils']
+           'fourier_transform', 'integrate']
