+++ qspectra/dynamics/generic.py
