+++ qspectra/hamiltonian.py
@@ -8,7 +8,7 @@ from operator_tools import (transition_operator, operator_extend, unit_vec,
                             tensor, extend_vib_operator, vib_create,
                             vib_annihilate)
 from polarization import polarization_vector, random_rotation_matrix
-from utils import imemoize, memoized_property, Zero
+from utils import imemoize, memoized_property, ZeroArray
 
 
 class HamiltonianError(Exception):
@@ -447,7 +447,7 @@ def optional_ensemble_average(func):
             random_seed = kwargs.pop('ensemble_random_seed', None)
             random_orientations = kwargs.pop(
                 'ensemble_random_orientations', False)
-            total_signal = Zero()
+            total_signal = ZeroArray()
             for dyn_model in dynamical_model.sample_ensemble(
                     ensemble_size, random_orientations, random_seed):
                 (t, signal) = func(dyn_model, *args, **kwargs)
