+++ qspectra/polarization.py
@@ -5,7 +5,7 @@ from numpy import cos, sin, pi, sqrt
 import inspect
 import numpy as np
 
-from .utils import Zero
+from .utils import ZeroArray
 
 
 COORD_MAP = {'x': np.array([1, 0, 0]),
@@ -143,7 +143,7 @@ def optional_4th_order_isotropic_average(func):
             kwargs = _get_call_args(func, *args, **kwargs)
             weights = invariant_weights_4th_order(kwargs.pop('polarization'))
             signals = {}
-            total_signal = Zero()
+            total_signal = ZeroArray()
             for invariant, weight in zip(FOURTH_ORDER_INVARIANTS, weights):
                 if weight > 1e-8:
                     for p in list_polarizations(invariant):
@@ -174,7 +174,7 @@ def optional_2nd_order_isotropic_average(func):
             kwargs = _get_call_args(func, *args, **kwargs)
             weight = np.dot(*map(polarization_vector,
                                  kwargs.pop('polarization')))
-            total_signal = Zero()
+            total_signal = ZeroArray()
             for p in ['xx', 'yy', 'zz']:
                 (t, signal) = func(polarization=p, **kwargs)
                 total_signal += signal
